+++ lib/src/types/main.dart
@@ -215,3 +215,6 @@ export 'printer.dart' show Printer;
 export 'window_type.dart' show WindowType;
 export 'window_style_mask.dart' show WindowStyleMask;
 export 'window_titlebar_separator_style.dart' show WindowTitlebarSeparatorStyle;
+export 'custom_tabs_navigation_event_type.dart' show CustomTabsNavigationEventType;
+export 'custom_tabs_relation_type.dart' show CustomTabsRelationType;
+export 'prewarming_token.dart' show PrewarmingToken;